requirements.txt
streamlit==1.26.0
tensorflow==2.15.0
numpy==1.24.3
imageio==2.31.1
opencv-python==4.7.0.72
matplotlib==3.7.1
ffmpeg-python==0.2.0
Pillow==9.5.0
